+++ dexp/processing/restoration/test/test_dehaze.py
@@ -1,7 +1,8 @@
+import numpy
+
 from dexp.processing.backends.cupy_backend import CupyBackend
 from dexp.processing.backends.numpy_backend import NumpyBackend
 from dexp.processing.restoration.dehazing import dehaze
-
 from dexp.processing.synthetic_datasets.nuclei_background_data import generate_nuclei_background_data
 from dexp.utils.timeit import timeit
 
@@ -27,7 +28,8 @@ def _test_dehaze(backend, length_xy=128):
                                                                       add_noise=True,
                                                                       length_xy=length_xy,
                                                                       length_z_factor=4,
-                                                                      independent_haze=True)
+                                                                      independent_haze=True,
+                                                                      dtype=numpy.float32)
 
     with timeit('dehaze_new'):
         dehazed = dehaze(backend, image, size=25, in_place=False)
